AIOEC-1698
Back to use int instead of boolean during json response.
Added params to toggleClass

+++ lib/twig/cache.php
@@ -23,7 +23,7 @@ class Ai1ec_Twig_Cache extends Ai1ec_Base {
 				'http.response.render.strategy.json'
 		);
 		$output['data'] = array(
-			'state' => false !== $cache_dir,
+			'state' => (int)(false !== $cache_dir),
 		);
 		$render_json->render( $output );
 	}
